An update using this prompt:
```I found a problem, and I think we now need to define it.  For gregorian dates before 08/18/1989 we get an output like this:
Converted Date in Philomena:
Year -1, Month -1 (undefined), Day -1
Time at midnight in PUJT: 05:00 PUJT-5

So I think we need to have a date system handle before this date.  In the gregorian calendar we count dates with BC meaning "Before Christ".  For the philomena calendar we can also use BC and have it mean "Before Cunk".  Can you fix the code to adjust for this?  Also, can you add to the descriptive text in the middle what our new BC definition is?```